% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpMX.R
\name{combine_hexp}
\alias{combine_hexp}
\title{Combine Homeolog Expression Across Subgenomes}
\usage{
combine_hexp(x, subgenomes, name_to = NULL)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object.}

\item{subgenomes}{A vector of indices specifying which subgenomes to combine.}

\item{name_to}{A character string specifying the name of the new combined
subgenome. If not provided, the name will be generated by concatenating
the names of the combined subgenomes.}
}
\value{
An updated \linkS4class{ExpMX} object with the specified subgenomes
combined.
}
\description{
Sums homeolog expression values from multiple subgenomes into a single
subgenome.
}
\details{
An \code{ExpMX} object stores homeolog expression data for each subgenome as a
list of matrices in the \code{data} slot. This function aggregates selected
subgenome expression matrices into one combined matrix and updates the
relevant fields in the \code{ExpMX} object accordingly.

This function is particularly useful for analyzing homeolog expression in
species like wheat, which has three subgenomes (commonly named A, B, and D).
For example, to analyze the expression ratio between the combined expression
of subgenomes A and B versus D, users can apply this function to merge
A and B into a single subgenome.
}
\examples{
x <- sim_homeolog_counts(n_genes = 100, n_subgenomes = 3)
x_combined <- combine_hexp(x, subgenomes = c(1, 2))
x_combined

}
