% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{get_sim_params}
\alias{get_sim_params}
\title{Retrieve Simulation Parameters}
\usage{
get_sim_params(x, param = NULL)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object
generated by \code{sim_homeolog_counts()}.}

\item{param}{Character. A character string specifying which parameter
to retrieve. Valid options include \code{"mu"} (true means),
\code{"her"} (homeolog expression ratios),
\code{"params"} (sampled means and dispersions),
and \code{"params_population"} (population-level values).}
}
\value{
A list of \linkS4class{DataFrame} objects
containing the requested parameter values.
}
\description{
Retrieves the mean and dispersion values used to simulate read count data
from a negative binomial distribution.
}
\details{
The \code{sim_homeolog_counts()} function generates artificial read counts
for each homeolog using a negative binomial distribution.
The mean and dispersion values used for this simulation are derived
from real RNA-Seq datasets and stored in a \linkS4class{SimParams} object,
which is embedded within the returned \linkS4class{ExpMX} object.

This function provides an interface to extract specific simulation parameters
from the \linkS4class{SimParams} object.
}
\examples{
x <- sim_homeolog_counts()
param_mu <- get_sim_params(x, 'mu')
head(param_mu)

}
\seealso{
\code{\link[=sim_homeolog_counts]{sim_homeolog_counts()}}, \linkS4class{ExpMX},
\linkS4class{SimParams}
}
