% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{hespresso-package}
\alias{hespresso}
\alias{hespresso-package}
\title{hespresso: LRT for Detecting Shifts in Homeolog Expression Ratios from Allopolyploid RNA-seq Data}
\description{
hespresso blends homeolog expression with espresso, capturing the essence of power, energy, and focus in homeolog expression analysis. It provides statistical tests for detecting shifts in homeolog expression ratios among subgenomes of allopolyploid species across diverse conditions, using RNA-seq read count data. Take a sip of hespresso, and start your analysis strong!
}
\details{
For a comprehensive introduction to the package,
including a quick start and case studies,
see the package documentation site:
\url{https://bitdessin.github.io/hespresso}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bitdessin/hespresso}
  \item Report bugs at \url{https://github.com/bitdessin/hespresso/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@bitdessin.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

}
\keyword{internal}
