% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeoroq.R
\name{homeoroq}
\alias{homeoroq}
\title{HomeoRoq: Detecting Shifts in Homeolog Expression Ratios between Two Groups}
\usage{
homeoroq(
  x,
  chains = 4,
  iter_sampling = 10000,
  n_threads = getOption("mc.cores", 1)
)
}
\arguments{
\item{x}{An \linkS4class{SeqCountData} object containing normalized homeolog
expression data (i.e., RNA-seq read counts).}

\item{chains}{Integer. Specifies the number of independent MCMC chains used
in the statistical test.}

\item{iter_sampling}{Integer. Specifies the number of post-warmup samples
to draw per chain for estimating p-values.}

\item{n_threads}{Integer. Specifies the number of threads to use for parallel
processing across sampling iterations.
The default value is taken from the global option \code{'mc.cores'},
which can be set at the beginning of an R session as
\code{options(mc.cores = 4)}. If \code{'mc.cores'} is not set, the default is 1.}
}
\value{
A data.frame with one row per homeolog,
containing the following columns:
\itemize{
\item \code{pvalue}: p-value from the statistical test.
\item \code{qvalue}: Adjusted p-value using the Benjamini-Hochberg method.
\item \verb{sumexp__*__$}: Total read counts for subgenome \code{$} under condition \code{*}.
\item \verb{ratio__*__$}: Homeolog expression ratio for subgenome \code{$} under condition \code{*}.
\item \code{ratio_sd}: Standard deviation of homeolog expression ratios calculated from observed counts.
}
All statistics are computed directly from the observed read counts.
Additionally, \code{NA} in the output indicates
that the homeolog is expressed in only one condition,
making ratio comparisons infeasible.
}
\description{
A statistical method for detecting shifts in homeolog expression ratios
between two subgenomes in allopolyploids across two groups.
}
\details{
HomeoRoq estimates the probability that homeolog expression ratios
remain constant between two groups using RNA-seq read count data for
each homeolog. Since this probability cannot be calculated analytically,
HomeoRoq uses a Bayesian framework and performs Markov chain Monte Carlo (MCMC)
sampling to estimate the distribution of homeolog expression ratios
under the null hypothesis of no change between conditions.

The \code{homeoroq()} function provides a comprehensive interface
for detecting shifts in homeolog expression ratios across conditions.
By default, it performs 10,000 sampling iterations
per MCMC chain (\code{iter_sampling = 1e4}) across four chains (\code{chains = 4}).
However, for robust and reproducible estimation of statistical significance
(i.e., \emph{p}-values), it is recommended to conduct multiple independent sampling runs,
each with a sufficiently large number of iterations.
To replicate the settings used in the original HomeoRoq implementation
(Akama et al., NAR, 2014), set \code{chains = 10} and \code{iter_sampling = 1e4}.
}
\examples{
x <- sim_homeolog_counts(100)
x_output <- homeoroq(x)

# sampling options
x_output <- homeoroq(x, chains = 4, iter_sampling = 1000, n_threads = 4)

}
\references{
Akama S, Shimizu-Inatsugi R, Shimizu KK, Sese J.
Genome-wide quantification of homeolog expression ratio revealed
nonstochastic gene regulation in synthetic allopolyploid Arabidopsis.
Nucleic Acids Res. 42(6):e46, 2014. \doi{10.1093/nar/gkt1376}
}
