% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpMX.R
\name{newExpMX}
\alias{newExpMX}
\title{Load Expression Dataset}
\usage{
newExpMX(x, group, mapping_table)
}
\arguments{
\item{x}{A matrix or data frame of gene expression data,
where each column corresponds to a sample
and each row corresponds to a gene.
The row names should be gene names which match those in the \code{mapping_table}.}

\item{group}{A vector or a data frame describing the experimental design.
If a data frame, it must include a column named \code{group}
indicating sample groupings.}

\item{mapping_table}{A data frame
representing the mapping table of homeolog names.
Each column should correspond to a progenitor genome.}
}
\value{
An \linkS4class{ExpMX} class object containing homeolog expression data.
}
\description{
Organize gene expression data into homeolog expression matrices and store
them as an \code{ExpMX} class object.
}
\details{
This function arranges the given gene expression matrix
into homeolog expression matrices using a mapping table (\code{mapping_table}),
and stores the result as an \linkS4class{ExpMX} class object for downstream analyses.
}
\examples{
gexp <- read.table(system.file(package = 'hespresso', 'extdata', 'C_flexuosa.tsv.gz'),
                   header = TRUE, sep = '\t', row.names = 1)
group <- c('wet', 'wet', 'wet', 'dry', 'dry', 'dry')
hnames <- read.table(system.file(package = 'hespresso', 'extdata', 'C_flexuosa.homeolog.tsv.gz'),
                     header = TRUE, sep = '\t')
expmx <- newExpMX(gexp, group, hnames)

}
\seealso{
\linkS4class{ExpMX}
}
