% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_counts}
\alias{norm_counts}
\title{Normalize RNA-Seq Read Count Data}
\usage{
norm_counts(x, method = c("tmm", "med", "cpm"), round = FALSE)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object.}

\item{method}{Character. Specifies the normalization method to use.
Supported methods are
\code{'tmm'} (TMM: trimmed mean of M-values; used by \pkg{edgeR}),
\code{'med'} (median ratio method; used by \pkg{DESeq2}),
and \code{'cpm'} (CPM: counts per million).
Note that CPM is generally used for visualizations and
is \strong{NOT RECOMMENDED} for statistical analysis,
as it may distort variance structure.}

\item{round}{Logical. If \code{TRUE}, rounds normalized expression values
to integers using the \code{round()} function.}
}
\value{
An \linkS4class{ExpMX} object with normalized read counts.
}
\description{
Normalizes RNA-Seq read count data.
}
\details{
This function first calculates gene expression by summing homeolog expression
across all subgenomes. Then, it normalizes the gene expression values using
the specified method. After normalization, it redistributes the normalized
gene-level expression back to homeologs according to the original homeolog
expression ratios.
}
\examples{
x <- sim_homeolog_counts(n_genes = 100)
x <- norm_counts(x)

}
