% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_HER_distr}
\alias{plot_HER_distr}
\title{Visualize distribution of homeolog expression ratios}
\usage{
plot_HER_distr(x, base = 1)
}
\arguments{
\item{x}{An \linkS4class{SeqCountData} object.}

\item{base}{Integer or character. An index or name specifying the subgenome
used to calculate homeolog expression ratios.}
}
\value{
A list of ggplot histogram objects, one for each group.
}
\description{
Visualizes the distribution of homeolog expression ratios using histograms.
}
\examples{
x <- sim_homeolog_counts(100)
plot_HER_distr(x)

}
