% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCountData.R
\docType{class}
\name{SeqCountData-class}
\alias{SeqCountData-class}
\title{Class to Store RNA-seq Read Count Data}
\value{
An object of class \linkS4class{SeqCountData}.
}
\description{
An S4 class for storing RNA-seq read count data, including gene or homeolog
expression, experimental design, and associated metadata.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of read count matrices, with one matrix per subgenome.}

\item{\code{gene_names}}{A character vector containing gene names.}

\item{\code{exp_design}}{A data frame describing the experimental design.
It must include a column named \code{group} indicating sample groupings.}

\item{\code{meta}}{A list containing additional metadata related to the experiment
or processing steps.}
}}

\seealso{
\code{\link[=newSeqCountData]{newSeqCountData()}}
}
