% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCountData.R
\name{show,SeqCountData-method}
\alias{show,SeqCountData-method}
\title{Print Contents of an \code{SeqCountData} Object}
\usage{
\S4method{show}{SeqCountData}(object)
}
\arguments{
\item{object}{An \linkS4class{SeqCountData} object.}
}
\value{
The input \linkS4class{SeqCountData} object, returned invisibly.
}
\description{
Prints the contents of an \code{SeqCountData} object to the console.
}
\examples{
x <- sim_homeolog_counts(100)
x

}
\seealso{
\linkS4class{SeqCountData}
}
