% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimParams.R
\docType{class}
\name{SimParams-class}
\alias{SimParams-class}
\title{Class to Store Simulation Parameters}
\value{
An object of class \code{SimParams}.
}
\description{
An S4 class for storing simulation parameters
used to generate synthetic RNA-seq count data.
}
\section{Slots}{

\describe{
\item{\code{n_subgenomes}}{An integer representing the number of subgenomes.}

\item{\code{n_genes}}{An integer representing the number of genes or homeologs.}

\item{\code{n_groups}}{An integer representing the number of experimental groups.}

\item{\code{n_replicates}}{An integer representing the number of replicates per group.}

\item{\code{params_population}}{A data frame containing population-level mean
and dispersion values used for sampling.}

\item{\code{params}}{A data frame containing sampled mean and dispersion values
used to generate artificial count data.}

\item{\code{nls}}{An object generated by \code{nls()}
for modeling dispersion or variance as a function of the mean.}

\item{\code{mu}}{A matrix of true mean expression values
of the artificial count data.}

\item{\code{her}}{A matrix of true homeolog expression ratios
of the artificial count data.}
}}

\seealso{
\code{\link[=sim_homeolog_counts]{sim_homeolog_counts()}}
}
