% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpMX.R
\docType{methods}
\name{subset-ExpMX}
\alias{subset-ExpMX}
\alias{[,ExpMX-method}
\alias{[,ExpMX,ANY,ANY,ANY-method}
\title{Extract a Subset of an ExpMX Object}
\usage{
\S4method{[}{ExpMX,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} object.}

\item{i}{A logical or integer vector specifying the homeolog tuples
to retain.}

\item{j}{A logical or integer vector specifying the samples to retain.}

\item{...}{Unused. Included for consistency with the \code{[} generic.}

\item{drop}{Unused. Included for consistency with the \code{[} generic.}
}
\value{
A subsetted \linkS4class{ExpMX} object.
}
\description{
Extracts homeolog tuples or samples from an \code{ExpMX} object.
}
\examples{
x <- sim_homeolog_counts(n_genes = 100)
x_10 <- x[seq_len(10), ]

}
\seealso{
\linkS4class{ExpMX}
}
