% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{hespresso-package}
\alias{hespresso}
\alias{hespresso-package}
\title{hespresso: Statistical Test for Detecting Shifts in Homeolog Expression Ratios in Allopolyploids}
\description{
hespresso blends homeolog expression with espresso, capturing the essence of power, energy, and focus in homeolog expression analysis. It provides statistical tests to identify homeologs with differential expression ratios among subgenomes in allopolyploid species across diverse conditions, using RNA-Seq read count data. Take a sip of hespresso, and start your analysis strong!
}
\details{
For a comprehensive introduction to the package,
including a quick start and detailed examples, see the package vignettes.
}
\examples{
browseVignettes("hespresso")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bitdessin/hespresso}
  \item Report bugs at \url{https://github.com/bitdessin/hespresso/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@bitdessin.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

}
