% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobit.R
\name{hobit}
\alias{hobit}
\title{HOBIT: Detecting Shifts in Homeolog Expression Ratios}
\usage{
hobit(
  x,
  use_Dirichlet = FALSE,
  no_replicate = FALSE,
  eps = 0.001,
  dist = c("NB", "ZINB"),
  n_threads = getOption("mc.cores", 1),
  parallel_chains = 1,
  ...
)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object containing normalized homeolog
expression data (i.e., RNA-Seq read counts).}

\item{use_Dirichlet}{Logical. Whether to apply a Dirichlet prior distribution
for sampling HERs.}

\item{no_replicate}{Logical. Whether to treat all replicates as belonging to
a single condition when estimating dispersion. Set to \code{TRUE} if the
dataset lacks replicates to avoid errors during dispersion estimation.}

\item{eps}{Numeric. A small value specifying the minimum threshold for
homeolog expression. Expression values below this threshold are replaced
to avoid issues during MCMC sampling, particularly when drawing from a
NB distribution with near-zero means.}

\item{dist}{Character. The distribution to use for modeling expression data.
Options are \code{'NB'} and \code{'ZINB'} (Zero-Inflated NB).}

\item{n_threads}{Integer. Specifies the number of threads used for parallel
processing of homeologs. This parallelization is applied at the homeolog
level and is independent of the MCMC sampling process, which is
controlled by the \code{parallel_chains} argument.
The default value is taken from the global option \code{'mc.cores'},
which can be set at the beginning of an R session using
\code{options(mc.cores = 4)}. If \code{'mc.cores'} is not set, the default is 1.
In general, setting a higher value for \code{n_threads} while keeping
\code{parallel_chains = 1} often results in better performance compared
to allocating more threads to \code{parallel_chains}.}

\item{parallel_chains}{Integer. Specifies the number of threads used to
parallelize MCMC sampling across multiple chains. This argument is
passed directly to the \code{\link[cmdstanr]{sample}} function in the
\pkg{cmdstanr} package.
If both \code{n_threads} and \code{parallel_chains} are greater than 1, the total
number of threads required will be \verb{n_threads × parallel_chains}.
Make sure sufficient computational resources are available
to support this configuration.}

\item{...}{Additional arguments passed to the \code{\link[cmdstanr]{sample}}
function in the \pkg{cmdstanr} package
(e.g., \code{chains}, \code{iter_warmup}, \code{iter_sampling}, \code{thin}).}
}
\value{
A data.frame with one row per homeolog,
containing the following columns:
\itemize{
\item \code{pvalue}: p-value from the LRT using normalized likelihoods.
\item \code{qvalue}: Adjusted p-value using the Benjamini-Hochberg method.
\item \code{raw_pvalue}: p-value from the LRT using raw likelihoods.
\item \code{raw_qvalue}: Adjusted raw p-value using the Benjamini-Hochberg method.
\item \verb{theta0__*}: Posterior HER estimates shared across all conditions, where \code{*} denotes the subgenome name.
\item \verb{theta__*__$}: Posterior HER estimates specific to each condition, where \code{*} denotes the subgenome name and \code{$} denotes the group name.
\item \code{Dmax}: Maximum absolute difference in HERs between groups.
\item \code{ORmax}: Maximum odds ratio of HERs between groups.
\item \code{logLik_H1}: Log-likelihood of the full model.
\item \code{logLik_H0}: Log-likelihood of the reduced model.
}
}
\description{
A statistical method for detecting shifts in homeolog expressio ratios
in allopolyploids across diverse conditions.
}
\details{
HOBIT is a statistical test for detecting homeologs with differential
homeolog expression ratios (HERs) across multiple experimental conditions
in allopolyploid species, using RNA-Seq read count data.
It employs a likelihood ratio test (LRT) to compare two hierarchical models:
a full model that allows HERs to vary among conditions, and a reduced model
that assumes constant ratios across all conditions.

HOBIT models homeolog expression using an negative binomial (NB) distribution.
Specifically, the expression level of a homeolog on the \eqn{i}-th subgenome,
denoted by \eqn{x_{i}}, is assumed to follow an NB distribution with mean
\eqn{\mu\theta_{i}} and dispersion \eqn{\phi}, where \eqn{\mu} is total gene
expression across all subgenomes, \eqn{\theta_{i}} is HER for the \eqn{i}-th
subgenome, and \eqn{\phi} is dispersion parameter.

\deqn{
x_{i} \sim NB(\mu\theta_{i}, \phi)
}

By default, HOBIT samples \eqn{\mu} and \eqn{\boldsymbol{\theta}} from
uninformative prior distributions. The dispersion \eqn{\phi} is estimated
using the \code{\link[edgeR]{estimateDisp}} function from the \pkg{edgeR}
package.

With these assumptions, HOBIT performs Markov chain Monte Carlo (MCMC)
sampling to fit both full and reduced model and compute likelihoods between
the two models. The \code{hobit()} function provides an end-to-end interface for
parameter estimation, model fitting, likelihood computation, and the LRT.
}
\examples{
x <- sim_homeolog_counts(10)
x_output <- hobit(x)

# MCMC sampling options
x_output <- hobit(x, chains = 2, iter_warmup = 100, iter_sampling = 100)

# parallel processing
x_output <- hobit(x, n_threads = 1, parallel_chains = 8, iter_warmup = 100, iter_sampling = 100)
x_output <- hobit(x, n_threads = 8, parallel_chains = 1, iter_warmup = 100, iter_sampling = 100)

}
\references{
Sun J, Sese J, and Shimizu KK.
A moderated statistical test for detecting shifts in homeolog expression
ratios in allopolyploids.
bioRxiv 2025;2025.07.01.660977. \doi{10.1101/2025.07.01.660977}
}
