% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_HER}
\alias{plot_HER}
\title{Visualize Homeolog Expression Ratios between Two Groups}
\usage{
plot_HER(x, base = 1, groups = NULL, label = NULL, size = 3, alpha = 0.8)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object.}

\item{base}{Integer or character. An index or name specifying the subgenome
used for calculating homeolog expression ratios.}

\item{groups}{Character vector of length two specifying the group names
to compare.}

\item{label}{Optional character vector or factor indicating categories
or types of each homeolog. Points will be colored based on this grouping
if provided.}

\item{size}{Numeric. A value specifying the point size in the scatter plot.}

\item{alpha}{Numeric. A decimal number between 0 and 1 specifying the
transparency level of points in the scatter plot.}
}
\value{
A ggplot scatter chart object.
}
\description{
Creates a scatter plot to visualize homeolog expression ratios between two
groups.
}
\examples{
x <- sim_homeolog_counts(100)
plot_HER(x)

}
