% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_HER_distr}
\alias{plot_HER_distr}
\title{Visualize distribution of homeolog expression ratios}
\usage{
plot_HER_distr(x, base = 1)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object.}

\item{base}{Integer or character. An index or name specifying the subgenome
used to calculate homeolog expression ratios.}
}
\value{
A list of ggplot histogram objects, one for each group.
}
\description{
Visualizes distribution of homeolog expression ratio using histogram.
}
\examples{
x <- sim_homeolog_counts(100)
plot_HER_distr(x)

}
