% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ExpMX.R
\name{show,ExpMX-method}
\alias{show,ExpMX-method}
\title{Print Contents of an \code{ExpMX} Object}
\usage{
\S4method{show}{ExpMX}(object)
}
\arguments{
\item{object}{An \linkS4class{ExpMX} class object.}
}
\value{
The input \linkS4class{ExpMX} object, returned invisibly.
}
\description{
Prints the contents of an \code{ExpMX} object to the console.
}
\examples{
x <- sim_homeolog_counts(100)
x

}
\seealso{
\linkS4class{ExpMX}
}
