% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_homeolog_counts}
\alias{sim_homeolog_counts}
\title{Generate Artificial RNA-Seq Read Counts Simulating Homeolog Expression}
\usage{
sim_homeolog_counts(
  n_genes = 10000,
  n_replicates = c(3, 3),
  n_subgenomes = 2,
  group_names = NULL,
  subgenome_names = NULL,
  seed_expmx = NULL
)
}
\arguments{
\item{n_genes}{Integer. Number of genes to simulate.}

\item{n_replicates}{Integer vector. Number of replicates per group.}

\item{n_subgenomes}{Integer. Number of subgenomes.}

\item{group_names}{Character vector. Names of the condition groups.}

\item{subgenome_names}{Character vector. Names of the subgenomes.}

\item{seed_expmx}{A data.frame or matrix representing a normalized expression
matrix to be used for defining the sampling population.}
}
\value{
An object of class \linkS4class{ExpMX},
containing simulated expression data.
}
\description{
Simulates artificial RNA-Seq read counts for allopolyploids based on a
negative binomial distribution, mimicking homeolog expression patterns
across subgenomes.
}
\details{
This function performs the following steps to generate synthetic read counts:
(1) prepares a sampling population,
(2) modifies mean expression values to simulate differentially expressed genes,
(3) introduces variation in homeolog expression ratios,
(4) estimates dispersion parameters based on adjusted means, and
(5) samples counts from a negative binomial distribution using the calculated
means and dispersions.

By default, the function uses real RNA-Seq datasets from \emph{Cardamine flexuosa}
and wheat to define the sampling population for allopolyploids consiting of
two or three subgenomes, respectively, in step (1). Alternatively, users can
provide their own normalized expression matrix via the \code{seed_expmx} argument.
The matrix should have genes as rows and replicates (from the same condition)
as columns.
}
\examples{
x <- sim_homeolog_counts()

# simulating allotetraploids with 3 replicates in each group
x <- sim_homeolog_counts(n_genes = 100,
                         n_replicates = c(3, 3), n_subgenomes = 2)

# simulating allohexaploids with 5 replicats in each group
x <- sim_homeolog_counts(n_genes = 100,
                         n_replicates = c(5, 5), n_subgenomes =  3)
}
