% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpMX.R
\docType{methods}
\name{[-method}
\alias{[-method}
\alias{[,ExpMX-method}
\title{Extract a subset of ExpMX Object}
\usage{
\S4method{[}{ExpMX}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An \linkS4class{ExpMX} class object.}

\item{i}{A logical expression or an integer vector specifying which gene
(rows) to retain.}

\item{j}{A logical expression or an integer vector specifying which replicate
(columns) to retain.}

\item{...}{NOT USED. Included for consistency with the base \code{[} function.}

\item{drop}{NOT USED. Included for consistency with the base \code{[} function.}
}
\value{
A subsetted \linkS4class{ExpMX} class object.
}
\description{
Extracts a subset of homeologs or replicates from an \code{ExpMX} object.
}
\examples{
x <- sim_homeolog_counts(n_genes = 100)
x_10 <- x[seq_len(10), ]

}
\seealso{
\linkS4class{ExpMX}
}
